chanllenges: prepare for plugging multiple drivers alongside the nsupdate exec
